Increase HM01B0 pin drive strength when using 1-bit mode
On RedBoard RP2350, the D0 pin ends up connecting to the HSTX connector.
When an HDMI cable is plugged in, D0 connects to the CEC pin, which has
the effect of dramatically increasing the capacitance on the D0 pin,
resulting in a terrible eye pattern. Increasing the drive strength helps
clean it up a lot. Also probably beneficial to do this to the PCLK pin,
and on all boards when using 1-bit mode (highest PCLK frequency, same
as MCLK pin) to help maintain good signal quality.