Reduce OV5640 drive strength
Similar to the HM01B0, too high drive strength seems to result in
degraded signal integrity, which is not visible with an oscilloscope
due to the additional capacitance on the line. The default drive
strength was set to the max (4x), which was causing bytes to sometimes
be missed with the OV5640. Reducing to 2x seems to be much more stable
now, and still sufficiently strong to get a clean eye pattern on the
D0 pin while an HDMI cable is connected with the RP2350 RedBoard.